@@ -6,8 +6,10 @@
   "repository": "mozilla/voice-web",
   "main": "server/server.js",
   "dependencies": {
+    "@types/commonmark": "^0.27.0",
     "@types/copy-webpack-plugin": "^4.0.1",
     "@types/express": "^4.11.1",
+    "@types/express-session": "^1.15.8",
     "@types/extract-text-webpack-plugin": "^3.0.0",
     "@types/jest": "^22.0.1",
     "@types/lodash.debounce": "^4.0.3",
@@ -17,14 +19,17 @@
     "@types/node": "^9.3.0",
     "@types/node-fetch": "^1.6.7",
     "@types/node-static": "^0.7.0",
+    "@types/passport": "^0.4.3",
     "@types/random-js": "^1.0.30",
-    "@types/react": "^16.0.2",
+    "@types/react": "^16.1.0",
+    "@types/react-content-loader": "^2.0.1",
     "@types/react-dom": "^16.0.2",
     "@types/react-modal": "^3.1.1",
     "@types/react-redux": "^5.0.12",
     "@types/react-router": "^4.0.16",
     "@types/react-router-dom": "^4.2.0",
     "@types/redux-thunk": "^2.1.0",
+    "@types/request-promise-native": "^1.0.14",
     "@types/webpack": "^3.0.13",
     "@types/webpack-merge": "^4.1.0",
     "aws-sdk": "^2.67.0",
@@ -34,7 +39,9 @@
     "bcrypt": "^1.0.2",
     "bluebird": "^3.5.0",
     "body-parser": "^1.18.2",
+    "commonmark": "^0.28.1",
     "concurrently": "^3.5.0",
+    "cookie-parser": "^1.4.3",
     "css-loader": "^0.28.7",
     "cssnano": "^3.10.0",
     "db-migrate": "^0.10.2",
@@ -45,7 +52,9 @@
     "eslint": "^4.7.2",
     "eslint-plugin-prettier": "^2.3.1",
     "express": "^4.16.2",
+    "express-mysql-session": "^1.2.3",
     "express-promise-router": "^3.0.1",
+    "express-session": "^1.15.6",
     "extract-text-webpack-plugin": "^3.0.1",
     "ff": "0.2.2",
     "fluent": "^0.4.3",
@@ -64,22 +73,29 @@
     "node-fetch": "^1.7.3",
     "node-random-name": "^1.0.1",
     "node-static": "^0.7.10",
+    "passport": "^0.4.0",
+    "passport-auth0": "^0.6.1",
+    "passport-session": "^1.0.2",
     "postcss-cssnext": "^3.0.2",
+    "postcss-focus-within": "^2.0.0",
     "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.7",
     "prettier": "^1.7.0",
     "prom-client": "^10.1.1",
     "proxy-agent": "^2.0.0",
     "random-js": "^1.0.8",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^16.3.0",
+    "react-content-loader": "^3.1.1",
+    "react-dom": "^16.3.0",
     "react-modal": "^3.1.2",
-    "react-redux": "^5.0.6",
+    "react-redux": "^5.0.7",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "readline": "^1.3.0",
     "redux": "^3.7.2",
     "redux-thunk": "^2.2.0",
+    "request": "^2.85.0",
+    "request-promise-native": "^1.0.5",
     "reselect": "^3.0.1",
     "stream-transcoder": "0.0.5",
     "style-loader": "^0.19.0",
@@ -96,6 +112,7 @@
   "devDependencies": {
     "bundlesize": "^0.15.3",
     "chalk": "^2.3.0",
+    "jest-matcher-utils": "^22.4.0",
     "nodemon": "^1.12.1",
     "pre-commit": "^1.2.2",
     "react-dev-utils": "^4.2.1"
@@ -113,7 +130,7 @@
     "prettier": "yarn lint --fix",
     "start": "node scripts/dev-server.js",
     "start:prod": "node server/js/server.js",
-    "test": "jest && yarn build && bundlesize -f \"web/dist/*.js\" -s 110kb"
+    "test": "concurrently --kill-others-on-fail \"yarn build && bundlesize -f \"web/dist/*.js\" -s 115kb\" \"jest --forceExit\""
   },
   "pre-commit": [
     "prettier"
