@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Modal from '../modal/modal';
 import { Button, LabeledInput, LabeledTextArea } from '../ui/ui';
-const { Localized } = require('fluent-react');
+import { Localized } from 'fluent-react';
 
 interface Props {
   onRequestClose: () => void;
@@ -21,7 +21,7 @@ export default ({ onRequestClose }: Props) => (
 
       <br />
 
-      <Localized id="contact-form-email" attrs={{ label: true }}>
+      <Localized id="email-input" attrs={{ label: true }}>
         <LabeledInput label="Email" name="email" required type="text" />
       </Localized>
 
@@ -37,7 +37,7 @@ export default ({ onRequestClose }: Props) => (
         <Localized id="contact-required">
           <div />
         </Localized>
-        <Localized id="contact-submit">
+        <Localized id="submit-form-action">
           <Button type="submit" />
         </Localized>
         <div />
