@@ -1,3 +1,4 @@
+import { getConfig } from '../config-helper';
 import { config, S3 } from 'aws-sdk';
 
 if (process.env.HTTP_PROXY) {
@@ -10,7 +11,7 @@ if (process.env.HTTP_PROXY) {
 }
 
 export namespace AWS {
-  let s3 = new S3({ signatureVersion: 'v4' });
+  let s3 = new S3(getConfig().S3_CONFIG);
 
   export function getS3() {
     return s3;
