@@ -1,21 +1,47 @@
 import DB from './model/db';
-import { CommonVoiceConfig } from '../config-helper';
 import { DBClipWithVoters } from './model/db/tables/clip-table';
 import Cache from './cache';
+import {
+  IDEAL_SPLIT,
+  randomBucketFromDistribution,
+  rowsToDistribution,
+  Split,
+} from './model/split';
 
 /**
  * The Model loads all clip and user data into memory for quick access.
  */
 export default class Model {
-  config: CommonVoiceConfig;
-  db: DB;
+  db = new DB();
   clipCache = new Cache(count => this.db.findClipsWithFewVotes(count));
-  sentencesCache = new Cache(count => this.db.findSentencesWithFewClips(count));
+  sentencesCaches: { [bucket: string]: Cache<string> } = Object.keys(
+    IDEAL_SPLIT
+  ).reduce(
+    (obj, bucket) => ({
+      ...obj,
+      [bucket]: new Cache(count =>
+        this.db.findSentencesWithFewClips(bucket, count)
+      ),
+    }),
+    {}
+  );
+  clipDistribution: Split = {
+    train: 0,
+    dev: 0,
+    test: 0,
+  };
 
-  constructor(config: CommonVoiceConfig) {
-    this.config = config;
-    this.db = new DB(this.config);
+  constructor() {
+    this.cacheClipDistribution().catch((e: any) => {
+      console.error(e);
+    });
   }
+
+  cacheClipDistribution = async () => {
+    this.clipDistribution = rowsToDistribution(
+      await this.db.getClipBucketCounts()
+    );
+  };
 
   /**
    * Fetch a random clip but make sure it's not the user's.
@@ -30,40 +56,23 @@ export default class Model {
     );
   }
 
-  async findEligibleSentences(count: number): Promise<string[]> {
-    return this.sentencesCache.take(count);
-  }
-
-  private print(...args: any[]) {
-    args.unshift('MODEL --');
-    console.log.apply(console, args);
-  }
-
-  async printMetrics() {
-    const totalUserClients = await this.db.getClientCount();
-    const listeners = await this.db.getListenerCount();
-    const submitters = await this.db.getSubmitterCount();
-
-    const totalClips = await this.db.getClipCount();
-    const unverified = totalClips - (await this.db.getValidatedClipsCount());
-    const votes = await this.db.getVoteCount();
-
-    this.print(totalUserClients, ' total user clients');
-    this.print((listeners / totalUserClients).toFixed(2), '% users who listen');
-    this.print(
-      (submitters / totalUserClients).toFixed(2),
-      '% users who submit'
-    );
-    this.print(totalClips, ' total clips');
-    this.print(votes, ' total votes');
-    this.print(unverified, ' unverified clips');
+  async findEligibleSentences(
+    client_id: string,
+    count: number
+  ): Promise<string[]> {
+    const user = await this.db.getUserClient(client_id);
+    return this.sentencesCaches[user ? user.bucket : 'train'].take(count);
   }
 
   /**
    * Update current user
    */
   async syncUser(uid: string, data: any): Promise<void> {
-    return this.db.updateUser(uid, data);
+    return this.db.updateUser(
+      uid,
+      data,
+      randomBucketFromDistribution(this.clipDistribution)
+    );
   }
 
   /**
@@ -85,5 +94,12 @@ export default class Model {
    */
   cleanUp(): void {
     this.db.endConnection();
+  }
+
+  async saveClip(clipData: any) {
+    const clip = await this.db.saveClip(clipData);
+    if (clip) {
+      this.clipDistribution[clip.bucket]++;
+    }
   }
 }
