@@ -2,6 +2,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import promisify from '../../../promisify';
 import { hash } from '../../clip';
+import { randomBucketFromDistribution, rowsToDistribution } from '../split';
 
 const CWD = process.cwd();
 const SENTENCE_FOLDER = path.resolve(CWD, 'server/data/');
@@ -109,13 +110,34 @@ export async function importSentences(pool: any) {
   await pool.query(
     'DELETE FROM sentences WHERE id NOT IN (SELECT original_sentence_id FROM clips)'
   );
+
   await pool.query('UPDATE sentences SET is_used = FALSE');
 
+  const [rows] = await pool.query(
+    'SELECT bucket, COUNT(bucket) AS count FROM sentences GROUP BY bucket'
+  );
+  const distribution = rowsToDistribution(rows);
+
   for (const sentence of await loadSentences(SENTENCE_FOLDER)) {
-    await pool.query(
-      'INSERT INTO sentences (id, text, is_used) VALUES (?, ?, TRUE) ON DUPLICATE KEY UPDATE is_used = TRUE',
-      [hash(sentence), sentence]
+    const id = hash(sentence);
+
+    const [[sentenceExists]] = await pool.query(
+      'SELECT 1 FROM sentences WHERE id = ?',
+      [id]
     );
+
+    if (sentenceExists) {
+      await pool.query('UPDATE sentences SET is_used = TRUE WHERE id = ?', [
+        id,
+      ]);
+    } else {
+      const bucket = randomBucketFromDistribution(distribution);
+      distribution[bucket]++;
+      await pool.query(
+        'INSERT INTO sentences (id, text, is_used, bucket) VALUES (?, ?, TRUE, ?)',
+        [id, sentence, bucket]
+      );
+    }
   }
 
   const [[{ count }]] = (await pool.query(
