@@ -1,5 +1,5 @@
 import * as http from 'http';
-import { CommonVoiceConfig } from '../config-helper';
+import { getConfig } from '../config-helper';
 import {
   collectDefaultMetrics,
   register,
@@ -11,19 +11,17 @@ import {
 collectDefaultMetrics({ timeout: 5000 });
 
 export default class Prometheus {
-  config: CommonVoiceConfig;
   registry: Registry;
   requests: Counter;
   clip_cnt: Counter;
   api_cnt: Counter;
   prometheus_cnt: Counter;
 
-  constructor(config: CommonVoiceConfig) {
-    this.config = config;
+  constructor() {
     this.registry = register;
 
     // Do not run prometheus endpoints on non prod site.
-    if (this.config.PROD) {
+    if (getConfig().PROD) {
       this.requests = new Counter({
         name: 'voice_requests',
         help: 'Total Requests Served',
