@@ -1,6 +1,10 @@
+import * as fs from 'fs';
 import * as http from 'http';
+import * as path from 'path';
+import * as commonmark from 'commonmark';
 import * as express from 'express';
 import { NextFunction, Request, Response } from 'express';
+import * as request from 'request-promise-native';
 import Model from './lib/model';
 import API from './lib/api';
 import Logger from './lib/logger';
@@ -11,47 +15,79 @@ import {
   APIError,
 } from './lib/utility';
 import { importSentences } from './lib/model/db/import-sentences';
-import { CommonVoiceConfig, getConfig } from './config-helper';
+import { getConfig } from './config-helper';
+import authRouter from './auth-router';
+import { router as adminRouter } from './admin';
 
-const CLIENT_PATH = '../../web';
+const FULL_CLIENT_PATH = path.join(__dirname, '../web');
 
-const CSP_HEADER = `default-src 'none'; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; img-src 'self' www.google-analytics.com; media-src data: blob: https://*.amazonaws.com https://*.amazon.com; script-src 'self' 'sha256-WpzorOw/T4TS/msLlrO6krn6LdCwAldXSATNewBTrNE=' https://www.google-analytics.com/analytics.js; font-src 'self' https://fonts.gstatic.com; connect-src 'self'`;
+const CSP_HEADER = [
+  `default-src 'none'`,
+  `style-src 'self' 'unsafe-inline' https://fonts.googleapis.com`,
+  `img-src 'self' www.google-analytics.com`,
+  `media-src data: blob: https://*.amazonaws.com https://*.amazon.com`,
+  // Note: we allow unsafe-eval locally for certain webpack functionality.
+  `script-src 'self' 'unsafe-eval' 'sha256-it/hVbE0ffRQjkt+hTb6/JM7wKrTSMEK4CHF4s42Zu8=' https://www.google-analytics.com/analytics.js https://pontoon.mozilla.org/pontoon.js`,
+  `font-src 'self' https://fonts.gstatic.com`,
+  `connect-src 'self' https://pontoon.mozilla.org/graphql`,
+].join(';');
 
 export default class Server {
-  config: CommonVoiceConfig;
   app: express.Application;
   server: http.Server;
   model: Model;
   api: API;
   logger: Logger;
   isLeader: boolean;
-  heartbeat: any;
 
-  constructor(config?: CommonVoiceConfig) {
-    this.config = config ? config : getConfig();
-    this.model = new Model(this.config);
-    this.api = new API(this.config, this.model);
-    this.logger = new Logger(this.config);
+  constructor(options?: { bundleCrossLocaleMessages: boolean }) {
+    options = { bundleCrossLocaleMessages: true, ...options };
+    this.model = new Model();
+    this.api = new API(this.model);
+    this.logger = new Logger();
     this.isLeader = null;
 
     // Make console.log output json.
-    if (this.config.PROD) {
+    if (getConfig().PROD) {
       this.logger.overrideConsole();
     }
 
     const app = (this.app = express());
 
+    app.use((request, response, next) => {
+      // redirect to omit trailing slashes
+      if (request.path.substr(-1) == '/' && request.path.length > 1) {
+        const query = request.url.slice(request.path.length);
+        response.redirect(301, request.path.slice(0, -1) + query);
+      } else {
+        next();
+      }
+    });
+
+    app.use(authRouter);
     app.use('/api/v1', this.api.getRouter());
+    app.use(adminRouter);
 
     const staticOptions = {
       setHeaders: (response: express.Response) => {
-        this.config.PROD && response.set('Content-Security-Policy', CSP_HEADER);
+        // Only use CSP locally. In production, Apache handles CSP headers.
+        // See path: nubis/puppet/web.pp
+        !getConfig().PROD &&
+          response.set('Content-Security-Policy', CSP_HEADER);
       },
     };
-    app.use(express.static(__dirname + CLIENT_PATH, staticOptions));
+
+    app.use(express.static(FULL_CLIENT_PATH, staticOptions));
+
+    if (options.bundleCrossLocaleMessages) {
+      this.setupCrossLocaleRoute();
+    }
+
+    this.setupPrivacyAndTermsRoutes();
+
     app.use(
-      '*',
-      express.static(__dirname + CLIENT_PATH + '/index.html', staticOptions)
+      /(.*)/,
+      express.static(FULL_CLIENT_PATH + '/index.html', staticOptions)
     );
 
     app.use(
@@ -73,6 +109,73 @@ export default class Server {
     );
   }
 
+  private setupCrossLocaleRoute() {
+    const localesPath = path.join(FULL_CLIENT_PATH, 'locales');
+    const crossLocaleMessages = fs
+      .readdirSync(localesPath)
+      .reduce((obj: any, locale: string) => {
+        const filePath = path.join(localesPath, locale, 'cross-locale.ftl');
+        if (fs.existsSync(filePath)) {
+          obj[locale] = fs.readFileSync(filePath, 'utf-8');
+        }
+        return obj;
+      }, {});
+
+    this.app.get('/cross-locale-messages.json', (request, response) => {
+      response.json(crossLocaleMessages);
+    });
+  }
+
+  // cache it for a day
+  documentCache: { [name: string]: { [locale: string]: string } } = {};
+  private async fetchDocument(name: string, locale: string) {
+    if (!this.documentCache[name]) this.documentCache[name] = {};
+
+    let textHTML = this.documentCache[name][locale];
+
+    if (!textHTML) {
+      const [status, text] = await request({
+        uri: `https://raw.githubusercontent.com/mozilla/legal-docs/master/Common_Voice_${name}/${locale}.md`,
+        resolveWithFullResponse: true,
+      })
+        .then((response: any) => [response.statusCode, response.body])
+        .catch(response => [response.statusCode, null]);
+      if (status >= 400 && status < 500) {
+        textHTML = (await Promise.all(
+          ['en', 'es-CL', 'fr', 'pt-BR', 'zh-TW'].map(locale =>
+            this.fetchDocument(name, locale)
+          )
+        )).join('<br>');
+      } else if (status < 300) {
+        textHTML = new commonmark.HtmlRenderer().render(
+          new commonmark.Parser().parse(
+            // There's a parseable datetime string in the legal documents, which we don't need to show
+            (text as string).replace(/{:\sdatetime=".*" }/, '')
+          )
+        );
+      }
+    }
+
+    this.documentCache[name][locale] = textHTML;
+
+    return textHTML;
+  }
+
+  private setupPrivacyAndTermsRoutes() {
+    this.app.get(
+      '/privacy/:locale.html',
+      async ({ params: { locale } }, response) => {
+        response.send(await this.fetchDocument('Privacy_Notice', locale));
+      }
+    );
+    this.app.get(
+      '/terms/:locale.html',
+      async ({ params: { locale } }, response) => {
+        response.send(await this.fetchDocument('Terms', locale));
+      }
+    );
+  }
+
   /**
    * Log application level messages in a common format.
    */
@@ -90,9 +193,10 @@ export default class Server {
     }
 
     try {
+      const config = getConfig();
       this.isLeader = await isLeaderServer(
-        this.config.ENVIRONMENT,
-        this.config.RELEASE_VERSION
+        config.ENVIRONMENT,
+        config.RELEASE_VERSION
       );
       this.print('leader', this.isLeader);
     } catch (err) {
@@ -112,6 +216,7 @@ export default class Server {
 
     try {
       await this.model.performMaintenance();
+      await this.model.db.migrateUserClientBuckets();
       if (doImport) {
         await importSentences(await this.model.db.mysql.createPool());
       }
@@ -127,7 +232,6 @@ export default class Server {
    * Kill the http server if it's running.
    */
   kill(): void {
-    clearInterval(this.heartbeat);
     if (this.server) {
       this.server.close();
       this.server = null;
@@ -140,7 +244,7 @@ export default class Server {
    */
   listen(): void {
     // Begin handling requests before clip list is loaded.
-    let port = this.config.SERVER_PORT;
+    let port = getConfig().SERVER_PORT;
     this.server = this.app.listen(port, () =>
       this.print(`listening at http://localhost:${port}`)
     );
@@ -155,13 +259,6 @@ export default class Server {
     } catch (err) {
       console.error('could not connect to db', err);
     }
-  }
-
-  startHeartbeat(): void {
-    clearInterval(this.heartbeat);
-    this.heartbeat = setInterval(() => {
-      this.model.printMetrics();
-    }, 30 * 60 * 1000); // 30 minutes
   }
 
   /**
@@ -180,8 +277,6 @@ export default class Server {
     if (isLeader) {
       await this.performMaintenance(options.doImport);
     }
-
-    this.startHeartbeat();
   }
 
   /**
