@@ -4,13 +4,14 @@ import fetch from 'node-fetch';
 import { AWS } from '../lib/aws';
 import Schema from '../lib/model/db/schema';
 import ServerHarness from './lib/server-harness';
+import { getConfig } from '../config-helper';
 
 let serverHarness: ServerHarness;
 let schema: Schema;
 
 beforeAll(async () => {
   serverHarness = new ServerHarness();
-  schema = new Schema(serverHarness.mysql, serverHarness.config);
+  schema = new Schema(serverHarness.mysql);
   await serverHarness.run();
 });
 
@@ -31,19 +32,15 @@ afterAll(async () => {
   'recording is uploaded and inserted into the db',
   async () => {
     const sentence = 'Wubba lubba dub dub!';
-    await serverHarness.server.api.clip.saveSentence(sentence);
-    await fetch(
-      `http://localhost:${serverHarness.config.SERVER_PORT}/api/v1/clips`,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'audio/ogg; codecs=opus4',
-          uid: 'wat',
-          sentence: encodeURIComponent(sentence),
-        },
-        body: fs.createReadStream(path.join(__dirname, 'test.ogg')),
-      }
-    );
+    await fetch(`http://localhost:${getConfig().SERVER_PORT}/api/v1/clips`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'audio/ogg; codecs=opus4',
+        uid: 'wat',
+        sentence: encodeURIComponent(sentence),
+      },
+      body: fs.createReadStream(path.join(__dirname, 'test.ogg')),
+    });
     expect(await serverHarness.getClipCount()).toBe(1);
   }
 );
