@@ -2,18 +2,18 @@ import * as React from 'react';
 import { Button } from './ui/ui';
 
 interface Props {
+  children?: any;
   disabled: boolean;
   percent: number;
-  text: string;
   onClick?(): void;
 }
 
-export default ({ disabled, percent, text, onClick }: Props) => (
+export default ({ disabled, percent, children, onClick }: Props) => (
   <Button className="progress-button" onClick={onClick} disabled={disabled}>
     <span
       className="progress"
       style={{ transform: `translateX(${percent - 100}%)` }}
     />
-    {text}
+    {children}
   </Button>
 );
