@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Link } from 'react-router-dom';
+import { LocaleLink } from '../locale-helpers';
 
 export const Button = ({
   className = '',
@@ -21,7 +21,7 @@ export const Button = ({
 
 export const CardAction = (props: any) =>
   props.to ? (
-    <Link className="card-action" {...props} />
+    <LocaleLink className="card-action" {...props} />
   ) : (
     <Button outline className="card-action" {...props} />
   );
@@ -34,17 +34,24 @@ const LabeledFormControl = ({
   label,
   required,
   ...props
-}: any) => (
-  <label
-    className={['labeled-form-control', 'for-' + Component, className].join(
-      ' '
-    )}
-    {...props}>
-    {label}
-    {required && '*'}
-    <Component {...{ required, ...props }} />
-  </label>
-);
+}: any) => {
+  const child = <Component {...{ required, ...props }} />;
+  return (
+    <label
+      className={['labeled-form-control', 'for-' + Component, className].join(
+        ' '
+      )}
+      {...props}>
+      {label}
+      {required && '*'}
+      {Component == 'select' ? (
+        <div className="wrapper with-down-arrow">{child}</div>
+      ) : (
+        child
+      )}
+    </label>
+  );
+};
 
 export const LabeledInput = (props: any) => (
   <LabeledFormControl component="input" {...props} />
