@@ -1,3 +1,4 @@
+import { Locale } from '../stores/locale';
 import { User } from '../stores/user';
 
 export interface Clip {
@@ -21,9 +22,11 @@ const CLIP_PATH = API_PATH + '/clips';
 const SENTENCES_PATH = API_PATH + '/sentences';
 
 export default class API {
-  private user: User.State;
+  private readonly locale: Locale.State;
+  private readonly user: User.State;
 
-  constructor(user: User.State) {
+  constructor(locale: Locale.State, user: User.State) {
+    this.locale = locale;
     this.user = user;
   }
 
@@ -42,10 +45,13 @@ export default class API {
             'Content-type': isJSON
               ? 'application/json; charset=utf-8'
               : 'text/plain',
-            uid: this.user.userId,
           },
           headers
         );
+
+        if (path.startsWith(location.origin)) {
+          finalHeaders.uid = this.user.userId;
+        }
 
         for (const header of Object.keys(finalHeaders)) {
           request.setRequestHeader(header, finalHeaders[header]);
@@ -128,10 +134,14 @@ export default class API {
     return this.fetch(CLIP_PATH + '/validated_hours');
   }
 
-  fetchLocale(locale: string): Promise<string> {
+  fetchLocaleMessages(locale: string): Promise<string> {
     return this.fetch(`/locales/${locale}/messages.ftl`, {
       isJSON: false,
     });
+  }
+
+  async fetchCrossLocaleMessages(): Promise<string[][]> {
+    return Object.entries(await this.fetch(`/cross-locale-messages.json`));
   }
 
   fetchRequestedLanguages(): Promise<string[]> {
@@ -142,6 +152,35 @@ export default class API {
     return this.fetch(`${API_PATH}/requested_languages`, {
       method: 'POST',
       body: { language },
+    });
+  }
+
+  fetchPontoonLanguages() {
+    return this.fetch('https://pontoon.mozilla.org/graphql', {
+      method: 'POST',
+      body: {
+        query: `{
+          project(slug: "common-voice") {
+            slug
+            localizations {
+              totalStrings
+              approvedStrings
+              locale {
+                code
+                name
+                population
+              }
+            }
+          }
+        }`,
+        variables: null,
+      },
+    });
+  }
+
+  fetchDocument(name: 'privacy' | 'terms'): Promise<string> {
+    return this.fetch(`/${name}/${this.locale}.html`, {
+      isJSON: false,
     });
   }
 }
