@@ -6,7 +6,8 @@ import { Recordings } from './recordings';
 import StateTree from './tree';
 import { User } from './user';
 import { Clips } from './clips';
-import { RequestedLanguages } from './requested-langauges';
+import { RequestedLanguages } from './requested-languages';
+import { Locale } from './locale';
 
 const USER_KEY = 'userdata';
 
@@ -23,18 +24,20 @@ try {
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 const store = createStore(
   function root(
-    { recordings, user, clips, requestedLanguages }: StateTree = {
+    { recordings, user, clips, requestedLanguages, locale }: StateTree = {
       api: undefined,
       recordings: undefined,
       user: undefined,
       clips: undefined,
       requestedLanguages: undefined,
+      locale: undefined,
     },
     action:
       | Recordings.Action
       | User.Action
       | Clips.Action
       | RequestedLanguages.Action
+      | Locale.Action
   ): StateTree {
     const newState = {
       recordings: Recordings.reducer(recordings, action as Recordings.Action),
@@ -44,15 +47,20 @@ const store = createStore(
         requestedLanguages,
         action as RequestedLanguages.Action
       ),
+      locale: Locale.reducer(locale, action as Locale.Action),
     };
 
-    return { api: new API(newState.user), ...newState };
+    return { api: new API(newState.locale, newState.user), ...newState };
   },
   preloadedState,
   composeEnhancers(applyMiddleware(thunk))
 );
 
-store.dispatch(User.actions.update({}));
+[
+  User.actions.update({}),
+  Clips.actions.refillCache(),
+  Recordings.actions.buildNewSentenceSet(),
+].forEach(store.dispatch.bind(store));
 
 const fieldTrackers: any = {
   email: () => trackProfile('give-email'),
