@@ -2,7 +2,7 @@ import API from '../services/api';
 import { Recordings } from './recordings';
 import { User } from './user';
 import { Clips } from './clips';
-import { RequestedLanguages } from './requested-langauges';
+import { RequestedLanguages } from './requested-languages';
 
 export default interface StateTree {
   api: API;
@@ -10,4 +10,5 @@ export default interface StateTree {
   user: User.State;
   clips: Clips.State;
   requestedLanguages: RequestedLanguages.State;
+  locale: string;
 };
